Fix member lookup for meetings.
The member lookup for meetings used to use an obscure tribunal api to lookup iso's(card swipes) and mnums. I've changed this query to be an internal database query. I deprecated the old function and noted the needed changes should we ever wish to return to an iso system. I also removed the card swipe field from the meeting page.